Convert the plugin to a package
