+++ .github/workflows/typeshed_primer.yml
@@ -3,7 +3,7 @@ name: typeshed_primer
 on:
   pull_request:
     paths:
-      - "pyi.py"
+      - "flake8_pyi/**/*"
       - ".github/**/*"
   workflow_dispatch:
