+++ flake8_pyi/__init__.py
@@ -1,0 +1,3 @@
+from .checker import PyiTreeChecker
+
+__all__ = ["PyiTreeChecker"]
